requirements.txt
streamlit
streamlit-extras
pypdf
openpyxl
langchain_community
langchain-openai
langchain
faiss-cpu